se modifica el valor double de weight por BigDecimal

+++ src/main/java/com/record/DeepDiveRecord/api/domain/diveday/getdivedaybyid/FishingResponse.java
@@ -2,6 +2,8 @@ package com.record.DeepDiveRecord.api.domain.diveday.getdivedaybyid;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -11,7 +13,7 @@ public class FishingResponse {
     private Integer id;
     private String notes;
     private boolean caught;
-    private double weight;
+    private BigDecimal weight;
     private double latG;
     private double longG;
     private String name;
